Initial commit for ComponentPool and Emitter classes.
Created a class for pooling components. In my tests so far, there is no
performance gain from simply newing and removing components.

Emitter is a class built on top of the ComponentPool that creates a
defined Component at a certain interval. Nice for particle systems.

Removed Level related stuff. Support was shoddy at best.

Fixed a few bugs with the Sprite class. The render method was using the
offeset incorrectly resulting in weird behaior. Also the Sprites image
now gets set on init when using an animation.

Fixed bug in the EventManager class. The list of subscribers is now a
weakref that cleans up on calls to notify().

Added some additional safeguards in the Component class.